+++ tests/unit/test_oclcClassify_process.py
@@ -1,13 +1,9 @@
 import datetime
-from lxml import etree
 import os
 import pytest
 
 from tests.helper import TestHelpers
 from processes import ClassifyProcess
-from managers import ClassifyManager
-from managers.oclcClassify import ClassifyError
-
 
 class TestOCLCClassifyProcess:
     @classmethod
@@ -28,20 +24,8 @@ class TestOCLCClassifyProcess:
                 self.rabbitQueue = os.environ['OCLC_QUEUE']
                 self.rabbitRoute = os.environ['OCLC_ROUTING_KEY']
                 self.oclc_catalog_manager = mocker.MagicMock()
-        
+
         return TestClassifyProcess('TestProcess', 'testFile', 'testDate')
-    
-    @pytest.fixture
-    def testXMLResponse(self):
-        def constructResponse(code, responseBlock):
-            return etree.fromstring('''<?xml version="1.0"?>
-                <classify xmlns="http://classify.oclc.org" xmlns:oclc="http://classify.oclc.org">
-                    <response code="{}"/>
-                    {}
-                </classify>
-            '''.format(code, responseBlock))
-        
-        return constructResponse
 
     @pytest.fixture
     def runProcessMocks(self, mocker):
@@ -194,7 +178,7 @@ class TestOCLCClassifyProcess:
         mockBulkSave.assert_called_once_with(['rec1', 'rec2', 'rec3'])
 
     def test_frbrizeRecord_success_valid_author(self, testInstance, testRecord, mocker):
-        mockIdentifiers = mocker.patch.object(ClassifyManager, 'getQueryableIdentifiers')
+        mockIdentifiers = mocker.patch.object(ClassifyProcess, '_get_queryable_identifiers')
         mockIdentifiers.return_value = ['1|test']
 
         mockRedisCheck = mocker.patch.object(ClassifyProcess, 'checkSetRedis')
@@ -209,7 +193,7 @@ class TestOCLCClassifyProcess:
         mockClassifyRec.assert_called_once_with('1', 'test', 'Author, Test', 'Test Record')
 
     def test_frbrizeRecord_success_author_missing(self, testInstance, testRecord, mocker):
-        mockIdentifiers = mocker.patch.object(ClassifyManager, 'getQueryableIdentifiers')
+        mockIdentifiers = mocker.patch.object(ClassifyProcess, '_get_queryable_identifiers')
         mockIdentifiers.return_value = ['1|test']
 
         mockRedisCheck = mocker.patch.object(ClassifyProcess, 'checkSetRedis')
@@ -225,7 +209,7 @@ class TestOCLCClassifyProcess:
         mockClassifyRec.assert_called_once_with('1', 'test', None, 'Test Record')
 
     def test_frbrizeRecord_identifier_in_redis(self, testInstance, testRecord, mocker):
-        mockIdentifiers = mocker.patch.object(ClassifyManager, 'getQueryableIdentifiers')
+        mockIdentifiers = mocker.patch.object(ClassifyProcess, '_get_queryable_identifiers')
         mockIdentifiers.return_value = ['1|test']
 
         mockRedisCheck = mocker.patch.object(ClassifyProcess, 'checkSetRedis')
@@ -241,7 +225,7 @@ class TestOCLCClassifyProcess:
         mockClassifyRec.assert_not_called
 
     def test_frbrizeRecord_identifier_missing(self, testInstance, testRecord, mocker):
-        mockIdentifiers = mocker.patch.object(ClassifyManager, 'getQueryableIdentifiers')
+        mockIdentifiers = mocker.patch.object(ClassifyProcess, '_get_queryable_identifiers')
         mockIdentifiers.return_value = []
 
         mockRedisCheck = mocker.patch.object(ClassifyProcess, 'checkSetRedis')
@@ -253,68 +237,6 @@ class TestOCLCClassifyProcess:
         mockIdentifiers.assert_called_once_with(testRecord.identifiers)
         mockRedisCheck.assert_not_called()
         mockClassifyRec.assert_called_once_with(None, None, 'Author, Test', 'Test Record')
-        
-    def test_classifyRecordByMetadata_success(self, testInstance, mocker):
-        mockClassifier = mocker.patch('processes.oclcClassify.ClassifyManager')
-        mockClassifierInstance = mocker.MagicMock(addlIds=[])
-        mockClassifier.return_value = mockClassifierInstance
-        mockClassifierInstance.getClassifyResponse.return_value = [
-            'xml1', 'xml2', 'xml3'
-        ]
-
-        mockCheckFetched = mocker.patch.object(ClassifyProcess, 'checkIfClassifyWorkFetched')
-        mockCheckFetched.side_effect = [False, True, False]
-        mockCreateDCDW = mocker.patch.object(ClassifyProcess, 'createClassifyDCDWRecord')
-
-        testInstance.classifyRecordByMetadata('1', 'test', 'testAuthor', 'testTitle')
-
-        mockClassifier.assert_called_once_with(
-            iden='1', idenType='test', author='testAuthor', title='testTitle'
-        )
-
-        mockClassifierInstance.getClassifyResponse.assert_called_once()
-
-        mockCheckFetched.assert_has_calls([
-            mocker.call('xml1'), mocker.call('xml2'), mocker.call('xml3')
-        ])
-
-        mockClassifierInstance.checkAndFetchAdditionalEditions.assert_has_calls([
-            mocker.call('xml1'), mocker.call('xml3')
-        ])
-
-        mockCreateDCDW.assert_has_calls([
-            mocker.call('xml1', [], '1', 'test'), mocker.call('xml3', [], '1', 'test')
-        ])
-
-    def test_classifyRecordByMetadata_error(self, testInstance, mocker):
-        mockClassifier = mocker.patch('processes.oclcClassify.ClassifyManager')
-        mockClassifierInstance = mocker.MagicMock()
-        mockClassifier.return_value = mockClassifierInstance
-        mockClassifierInstance.getClassifyResponse.side_effect = ClassifyError
-
-        mocker.patch.object(ClassifyProcess, 'createClassifyDCDWRecord')
-
-        with pytest.raises(ClassifyError):
-            testInstance.classifyRecordByMetadata('1', 'test', 'testAuthor', 'testTitle')
-
-    def test_createClassifyDCDWRecord(self, testInstance, mocker):
-        mockMappingInstance = mocker.MagicMock()
-        mockMappingInstance.record.identifiers = ['1|test', '2|test']
-        mockMapping = mocker.patch('processes.oclcClassify.ClassifyMapping')
-        mockMapping.return_value = mockMappingInstance
-
-        mockAddDCDW = mocker.patch.object(ClassifyProcess, 'addDCDWToUpdateList')
-        mockfetchOCLC = mocker.patch.object(ClassifyProcess, 'fetchOCLCCatalogRecords')
-
-        testInstance.createClassifyDCDWRecord('testXML', [], '1', 'test')
-
-        mockMapping.assert_called_once_with(
-            'testXML', {'oclc': 'http://classify.oclc.org'}, {}, ('1', 'test')
-        )
-        mockMappingInstance.applyMapping.assert_called_once()
-        mockMappingInstance.extendIdentifiers.assert_called_once_with([])
-        mockAddDCDW.assert_called_once_with(mockMappingInstance)
-        mockfetchOCLC.assert_called_once_with(['1|test', '2|test'])
 
     def test_fetchOCLCCatalogRecords_no_redis_match(self, testInstance, mocker):
         mockRedisCheck = mocker.patch.object(ClassifyProcess, 'multiCheckSetRedis')
@@ -349,12 +271,5 @@ class TestOCLCClassifyProcess:
             'test_oclc_queue', 'test_oclc_key', {'oclcNo': '1', 'owiNo': '1'}
         )
 
-    def test_checkIfClassifyWorkFetched(self, testInstance, testXMLResponse, mocker):
-        testXML = testXMLResponse(2, '<work owi="1"/>')
-
-        mockCheckRedis = mocker.patch.object(ClassifyProcess, 'checkSetRedis')
-        mockCheckRedis.return_value = False
-
-        assert testInstance.checkIfClassifyWorkFetched(testXML) == False
-
-        mockCheckRedis.assert_called_once_with('classifyWork', '1', 'owi')
+    def test_get_queryable_identifiers(self, testInstance):
+        assert testInstance._get_queryable_identifiers(['1|isbn', '2|test']) == ['1|isbn']